+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v1.7.0
+- Add CLI paramerter `--use-env` to automatically load environment variables as variables ([#17](https://github.com/qetza/replacetokens/issues/17)).
+
 ## v1.6.0
 - Add support for case-insensitive glob pattern matching ([#13](https://github.com/qetza/replacetokens/issues/13)).
 - Add support for dot paths in glob pattern matching ([#12](https://github.com/qetza/replacetokens/issues/12)).